% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tml3.R
\name{tml3}
\alias{tml3}
\title{Cross-validated Targeted Minimum Loss-based Estimation}
\usage{
tml3(
  data,
  trt,
  outcome,
  covar,
  id = NULL,
  outcome_type = c("binomial", "continuous"),
  folds = 10,
  learners_trt = "glm",
  learners_outcome = "glm",
  learners_cens = learners_trt,
  control = tml3_control()
)
}
\arguments{
\item{control}{}
}
\description{
Cross-validated Targeted Minimum Loss-based Estimation
}
