% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tml3_control.R
\name{tml3_control}
\alias{tml3_control}
\title{Title}
\usage{
tml3_control(
  .trt_folds = NULL,
  .outcome_folds = NULL,
  .cens_folds = NULL,
  .trim = 0.999,
  .discrete = TRUE
)
}
\arguments{
\item{.discrete}{}
}
\description{
Title
}
