% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tml3_control.R
\name{tml3_control}
\alias{tml3_control}
\title{Title}
\usage{
tml3_control(
  .learners_trt_folds = NULL,
  .learners_outcome_folds = NULL,
  .learners_cens_folds = NULL,
  .trim = 0.999,
  .discrete = TRUE,
  .info = FALSE
)
}
\arguments{
\item{.info}{}
}
\description{
Title
}
